requirements.txt
streamlit
requests
numpy
Pillow
matplotlib
streamlit-drawable-canvas
fastapi
uvicorn
tensorflow
opencv-python-headless
